+++ examples/bridges/huggingface_example.py
@@ -84,9 +84,9 @@ for _ in range(3):
 
     sample.meshes.add_tree(MeshToCGNS(mesh))
     sample.add_scalar("scalar", np.random.randn())
-    sample.add_field("node_field", np.random.rand(1, len(points)), location="Vertex")
+    sample.add_field("node_field", np.random.rand(len(points)), location="Vertex")
     sample.add_field(
-        "cell_field", np.random.rand(1, len(points)), location="CellCenter"
+        "cell_field", np.random.rand(len(triangles)), location="CellCenter"
     )
 
     dataset.add_sample(sample)